Cast rays from p0 to p1 always
We were using a version of bresenhams that just flipped the x0,y0 and
x1,y1 under certain conditions. This would mean that the ray would cast
from the tip of the ray, which was typically OOB, back to the player.

I have very little idea why the changes I made here work, but they seem
to, so I forge on.